# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Superconducting Circuits API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/circuit-definitions": {
      "get": {
        "tags": [
          "definition-authoring",
          "circuit-definitions"
        ],
        "summary": "List Circuit Definitions",
        "operationId": "list_circuit_definitions_circuit_definitions_get",
        "parameters": [
          {
            "name": "search_query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Query"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "created_at",
                "updated_at",
                "name"
              ],
              "type": "string",
              "default": "updated_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "definition-authoring",
          "circuit-definitions"
        ],
        "summary": "Create Circuit Definition",
        "operationId": "create_circuit_definition_circuit_definitions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/circuit-definitions/{definition_id}": {
      "get": {
        "tags": [
          "definition-authoring",
          "circuit-definitions"
        ],
        "summary": "Get Circuit Definition",
        "operationId": "get_circuit_definition_circuit_definitions__definition_id__get",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Definition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "definition-authoring",
          "circuit-definitions"
        ],
        "summary": "Update Circuit Definition",
        "operationId": "update_circuit_definition_circuit_definitions__definition_id__put",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Definition Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "definition-authoring",
          "circuit-definitions"
        ],
        "summary": "Delete Circuit Definition",
        "operationId": "delete_circuit_definition_circuit_definitions__definition_id__delete",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Definition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/circuit-definitions/{definition_id}/publish": {
      "post": {
        "tags": [
          "definition-authoring",
          "circuit-definitions"
        ],
        "summary": "Publish Circuit Definition",
        "operationId": "publish_circuit_definition_circuit_definitions__definition_id__publish_post",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Definition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/circuit-definitions/{definition_id}/clone": {
      "post": {
        "tags": [
          "definition-authoring",
          "circuit-definitions"
        ],
        "summary": "Clone Circuit Definition",
        "operationId": "clone_circuit_definition_circuit_definitions__definition_id__clone_post",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Definition Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {},
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}